UI and Prep
I fixed the canvas UI scaling, so your HUD and buttons should all look
fine and dandy when you are running, no matter the size.

This is mainly a backup because I am moving into the stage where I am
going to be doing major edits to the code. This is because I want to
implement spaceship upgrading.

Stay tuned ;)